Support Tuples for Layer Attribution (#216)
Summary:
This PR adds support for layer inputs / outputs being tuples for all layer attribution algorithms. This also makes corresponding updates to documentation for each modified algorithm.

For all layer attribution algorithms, if the layer input / output is a single tensor, a tensor of attributions is still returned (maintaining consistency with the existing behavior), and for multiple tensors, a tuple is returned. NoiseTunnel was currently using inputs to decide whether the output should be a tuple, that behavior has been modified to be consistent with the described behavior for Layer Attribution.

This PR also contains a bug fix for targets in LayerDeepLift, which was not using the given targets.
Pull Request resolved: https://github.com/pytorch/captum/pull/216

Differential Revision: D19267503

Pulled By: vivekmig

fbshipit-source-id: 0a684bebf2f312b2d7149cc4ffdf2c4a4d55024d