Enable DataParallel for DeepLift/LayerDeepLift/NeuronDeepLift and its SHAP variants (#190)
Summary:
This PR enables DataParallel models for DeepLift/LayerDeepLift/NeuronDeepLift and its SHAP variants.
This required to pass inputs and baselines as a tuple and concatenate those in the main module's forward_pre_hook. The output is then being chunked in the forward_hook of the same module.

In addition to that I also needed to extend `compute_layer_gradients_and_eval ` and `_forward_layer_eval_with_neuron_grads` and enable a custom function that in this case chunks the inputs of the selected `layer` and the outputs of the selected `neuron` in the layer and neuron variants.

I also enabled cross_max output for the maxpool functions.
Pull Request resolved: https://github.com/pytorch/captum/pull/190

Differential Revision: D19388900

Pulled By: NarineK

fbshipit-source-id: e6cda9a54c0deeb7883e8424626a923d7ada59a5