Adding isort (#310)
Summary:
This applies isort to alphabetize import ordering. We also add an isort check to CircleCI to ensure that import ordering is maintained in future PRs. This required adding an isort config that is compatible with black, which was found here: https://github.com/psf/black

This also adds a type: ignore for an issue caused by merging Insights type fixes with type hints for common.
Pull Request resolved: https://github.com/pytorch/captum/pull/310

Reviewed By: orionr

Differential Revision: D20206324

Pulled By: vivekmig

fbshipit-source-id: 48cce82d60a9a6f7384211e7aae2ffb4c0c85772