Select targets gpu fix (#317)
Summary:
Hey,
Here is my proposed fix for [https://github.com/pytorch/captum/issues/316](https://github.com/pytorch/captum/issues/316).
I've added a test case that uses the BaseGPUTest class and simply runs the saliency target tests again, with everything on the GPU. As I did not fully understand the `_target_batch_test_assert` function and what the requirements for the tests are, I simply copied the function for now. I assume there may be an obvious way to integrate these tests. :)

Another question is:
Do we want to enforce the move if we receive a target tensor, which is not on the right device?
We could print a User Warning similar to automatically setting require gradients as in gradient.py:33.
Or should the user be responsible to move the target tensor to the correct device?

Best regards,
Kai
Pull Request resolved: https://github.com/pytorch/captum/pull/317

Reviewed By: NarineK

Differential Revision: D20371529

Pulled By: vivekmig

fbshipit-source-id: 92461d358f589bf47487d6170fac6a54e1d83123