Add sensitivity (#360)
Summary:
Adding sensitivity....

The implementation is more generic and can work with various explanation methods. Attribution approaches are examples of it.

Moved some of the generic logic that was updating the kwargs in noise tunnel to `common.py` and I'm reusing that logic in the sensitivity method.

More details about the approach can be found here: https://arxiv.org/pdf/1901.09392.pdf

This implementation also has batching support, documentation and test cases.
Type hints and DataParallel tests will be added in a separate PR.
Pull Request resolved: https://github.com/pytorch/captum/pull/360

Reviewed By: edward-io

Differential Revision: D21374348

Pulled By: NarineK

fbshipit-source-id: f61fcdc81dfc5b9abb1be59199aa28780e9c97b6