Speed up pytorch_captum tests by moving calculation of torch.cuda.is_available() up to module level (#402)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/402

In this diff we attempt to prevent listing of the tests in `test_jit.py` and `test_data_parallel.py` if CUDA is not available by moving the CUDA check up to the module level. This way we don't have to discover these tests and then attempt to run them only for it to trigger this exact code path somewhere down the setup stage for each test case.

The main reason why this takes so long is that the computation for `torch.cuda.is_available()` requires the import of the module `torch._C` which we've profiled in the past to take anywhere between 15-20 seconds. To do this for over 1000 tests every single time we test `//pytorch/captum:attributions` is expensive (both timewise and resource wise).

 ---

Reviewed By: vivekmig

Differential Revision: D21966198

fbshipit-source-id: b3fb0c71c90e26ae854a7e8128259032caf520de