Assert grad_input and inputs in DeepLift and fix the layer attribution issue for MaxPool (#390)
Summary:
Related to the issue: https://github.com/pytorch/captum/issues/382 asserting `grad_inputs` and `inputs` to have the same shape. More description about the workaround and why the issue happens can be found in the description of the assert. The error occurs when we attribute to the outputs of the layer because of the input or output tensor returned in the forward hook.

Added `forward_hook_with_return_excl_modules` that contains the list of modules for which we don't want to have a return in the forward_hook. This is only used in DeepLift and can be used for any algorithm that attributes to maxpool and at the same time has a backward hook set on it.

Added test cases for layer and neuron use cases.
Pull Request resolved: https://github.com/pytorch/captum/pull/390

Reviewed By: edward-io

Differential Revision: D22197030

Pulled By: NarineK

fbshipit-source-id: e6cf712103900190f46c5c1e9051519f3eaa933f