Neuron Aggregation (#495)
Summary:
This adds support for neuron aggregation, neuron_selector can be a function which returns a custom aggregate of a layer's neurons for all neuron methods other than neuron conductance, which has dependence on output gradients. The neuron_index argument was renamed, and a deprecation decorator was added to provide a warning for usage of the old parameter as a keyword argument. This decorator can be removed prior to the 0.4.0 release.

Documentation of the new callable functionality has been added to NeuronDeepLift, this documentation will be propagated to other relevant methods after review.

Pull Request resolved: https://github.com/pytorch/captum/pull/495

Reviewed By: miguelmartin75

Differential Revision: D24346065

Pulled By: vivekmig

fbshipit-source-id: c3853e19256de4c8c32a8ff615965bf513a5cd22