Jacobian calculation w.r.t. model parameters for a batch of inputs
Summary:
Implements two different methods to calculate the Jacobian w.r.t parameters for a batch of inputs.

The first method is equivalent to computing a batch-wise forward pass (and loss if necessary), and then computing the gradient for each input in a for-loop.

The second method is fully vectorized, but relies on custom code and only supports linear and conv2d layers.

Reviewed By: NarineK

Differential Revision: D29405094

fbshipit-source-id: 68eba5ae2eb38021b59787acc6f67854da6bf653