Enhance error message for missing rule (#727)
Summary:
This PR provides additional information to an error message that is triggered when neither an explicit LRP rule is defined nor a default rule exists. The current error message is quite short and does not give enough information to the user on how to solve the problem as discussed in issue https://github.com/pytorch/captum/issues/723 .

Feel free to propose a different wording.

Pull Request resolved: https://github.com/pytorch/captum/pull/727

Reviewed By: aobo-y

Differential Revision: D30045186

Pulled By: vivekmig

fbshipit-source-id: 2dc7f7a29d014da12ebdc409d4abc676d5cbbc34