change TCAV to use captum/_utils/fb/av.py
Summary:
AV.load now returns a Dataset, not a DataLoader
Also, activations for a single layer and multiple concepts are no longer stored as a single DataLoader (which also generates labels).  Instead, they are stored as a list of Datasets (not DataLoaders).  This list is used to create a Dataset (not DataLoader) that also generates labels.  I think it is better to pass around Datasets for as long as possible, because they are more flexible.

Reviewed By: NarineK

Differential Revision: D30566560

fbshipit-source-id: 1731e1afa3a34b6d445863fae0dce6c1a8425bd5