refactor to use `_basic_computation` (#848)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/848

`TracInCP._influence`, `TracInCPFast._influence`, `TracInCPFastAmortized._get_intermediate_quantities`, `TracInCPFastRandProj.__init__`, and `TracInCPFastRandProj._get_intermediate_quantities` are refactored.  In particular, a computation that is performed on both the input batch and training batches is separated into a `_basic_computation` method.  This computation will also be re-used once `_self_influence` is implemented, so that is why it is good to separate it out to avoid code duplication.  Since no new functionality is added, the tests do not change.

Differential Revision: D33660472

fbshipit-source-id: 57243c557ccf50fc0d7eaf6f57917a200c84e202