Enable attribute_to_layer_input both for layer activations and layer attribution (#864)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/864

Currently, we cannot specify `attribute_to_layer_input` to LayerActivation. Since this flag has to be consistent both for attribution and layer activation, we can pass it through the constructor and use the same value for both layer activation and attribution.

Removed unnecessary docs, e.g.: (a, b) (2 dimensions) There is no need for (2 dimensions). It is already clear that it has 2 dimensions.

Also, simplified the computations based on the magnitude. The original one had an issue that if we had only one example and that example had positive TCAV score the magnitude was always 1.0  and that wasn't very useful.

Reviewed By: aobo-y

Differential Revision: D33908812

fbshipit-source-id: b0c9378df7bc2b3e291800b25fe9065357cfbd82