require reduction = "sum" (#880)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/880

- require reduction of loss_fn for TracInCP to be "sum" if sample_wise_grads_per_batch==True, by changing documentation and assertions
- require reduction of loss_fn for TracInCPFast and TracInCPFastRandProj to be "sum", always, by changing documentation and assertions

Although reduction can technically be "mean" for TracInCP, I thought it would be more simple on the user if we just told them it has to be "sum".  Otherwise, we allow "sum" and "mean" for TracInCP, but only "sum" for TracInCPFast and TracInCPFastRandProj, which could be confusing.  In the future, we can allow reduction to be "mean" for TracInCPFast and TracInCPFastRandProj.  Then, we can allow both reductions for all classes.

This diff replaces D34354958, which changed deleted files.

Reviewed By: NarineK

Differential Revision: D34533617

fbshipit-source-id: 4beeb941e4019addd81adc871e2f63b9e41dd1a3