allow "mean" reduction for TracInCPFast, TracInCPFastRandProj (#913)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/913

This adds support for the "reduction" of the loss function used by `TracInCPFast` and `TracInCPFastRandProj` to be "mean", not just "sum" (as before).  This is done by appropriate scaling in `_basic_computation_tracincp_fast`.  This scaling could alternatively been handled in `_jacobian_loss_wrt_inputs`, if this is a better option, can do that instead.  Tests are run for more parameters - basically if a reduction="sum" test was run, a reduction="mean" test is now added.

Reviewed By: vivekmig

Differential Revision: D34802879

fbshipit-source-id: d244134910562de66d299d889ef08b00e22fad63