update TracInCP tutorial to use `train_dataset` argument
Summary: The API of TracIn was changed so that for all implementations, the initialization argument for the training dataset is now called `train_dataset` instead of `influence_src_dataset`.  In this diff, the TracIn tutorial is updated to reflect this change.  It just involves changing the named argument in TracIn constructors.

Reviewed By: NarineK

Differential Revision: D38378555

fbshipit-source-id: dddb4e320b094223964cb8d048821e9aa45281fe