Improve version checking (#999)
Summary:
Without the packaging library, statements like: `"1.8.0" > "1.10.0"` will be equal to True, despite v1.10 being a later version that v1.8.0.

The `packaging` library will in some cases not be already installed on a user's device, so I've also added it the `setup.py` file. It's one of the core libraries from the Python Packaging Authority, but it's not included with the base Python installation: https://packaging.python.org/en/latest/key_projects/#pypa-projects

This wasn't an issue in https://github.com/pytorch/captum/pull/940 as one the libraries in dev install has `packaging` as a dependency. So, there's no error when the tests are using the `packaging` library.

Pull Request resolved: https://github.com/pytorch/captum/pull/999

Reviewed By: vivekmig

Differential Revision: D38693600

Pulled By: NarineK

fbshipit-source-id: a5ea5ef6f2ca175d60f1638072add2fea6d31091