change tracin progress test (#1007)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/1007

the tests for `test_tracin_show_progress` were failing we check that the progress reaches 100% X times, but sometimes, the progress reaches 100% more than X times.  As aobo-y pointed out, this is because tqdm will sometimes correct its estimate of it/s for the total iteration over progress, and print 100% an additional time, with the updated estimate of it/s.  The fix is to check that progress reaches 100% at least X times.  We don't check that it reaches 100% either X or X+1 times, because all we can reasonably assume of tqdm is that it may re-estimate it/s >= 0 times.  Note that this change is to correct a problem with tqdm, *not* `SimpleProgress`.

Reviewed By: NarineK

Differential Revision: D38443861

fbshipit-source-id: 3d6b8588380014e4e6e4cf8e0dfd5464c50ce7be