Add support for nested progressbar to tracincp even when tqdm is not available. (#1046)
Summary:
This Pull request add support for nested progressbar in SimpleProgress bar (i.e. when tqdm is not installed) by leveraging Python contexts (with statements) that is also implemented in tqdm. To keep `SimpleProgress` simple this is the behaviour:
- Add a new line per each update of the parent progress bar.
- Guarantee that each update for parent progress bar is refreshed.
- Do not write refresh at the ending of parent progress bar (to avoid duplicate lines).

This support is used now in TracInCp and TracInCpFast methods.
![Screen Shot 2022-10-11 at 12 43 23 PM](https://user-images.githubusercontent.com/3238673/195717879-1ffc3e4a-a8d4-4f4f-a661-4c11fd93252c.png)

Pull Request resolved: https://github.com/pytorch/captum/pull/1046

Reviewed By: aobo-y

Differential Revision: D40397776

Pulled By: cyrjano

fbshipit-source-id: 26316255296a50fc4c80a22658be386769877c5e