make clamp optional in STG's get_gate_values (#1070)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/1070

add a flag in `get_gate_values` indicating if clamping the gate values. As smoothed Bernoulli variables, gate values are clamped withn 0 and 1 by defautl.

But turning this off to get the raw means of the underneath distribution (e.g., conrete, gaussian) can be useful to differentiate the gates' relative importance when multiple gate values are beyond 0 or 1.

Reviewed By: cyrjano

Differential Revision: D41013197

fbshipit-source-id: fcbc923efea07e539444b1ad5658462ecfc00e57