TracIn FBLearner Flow bug TracInCPFastRandProj projection_dim mismatch (#1134)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/1134

For F416836189, F417066355 and other runs of FBLearner Flow for computing influential embeddings using TracInCPFastRandProj, there is a possible bug of the final embedding size being of dimension 60 when "projection_dim" passed in is of 100.  Its possible that this isn't a bug per-say, but its something that we should look into to either fix or possibly document better.

Just for reference I haven't seen this issue using TracInCPFastRandProj without passing in "projection_dim" explicitly ( which is effectively just TracInCPFast such as F418560294) or when using Arnoldi Influence ( such as F418927001).

**Update, This is not a bug, but rather the intended output of the method.  I've added a test with different configurations to further explain how the projection method works using TracInCPFastRandProj

Reviewed By: NarineK

Differential Revision: D45050510

fbshipit-source-id: ca1e436da4efa8ba0ca8724ebf902d790884b382