check the length in assertTensorTuplesAlmostEqual test when expected is tuple
Summary:
When `expected` is a tuple in `assertTensorhenTuplesAlmostEqual`, check if the `actual` has the same size before iterating the elements.

It should also check if `actual` is a tuple, but the existing code has heavily abused this func to compare `tensor` vs `tuple`, `list` vs `tuple`. Have to leave the fix to the future.

Reviewed By: skim9

Differential Revision: D46307992

fbshipit-source-id: 4bdc199a752c85e5c602cd12057082e356ce34ac