check the length in assertTensorTuplesAlmostEqual test when expected is tuple (#1144)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/1144

When `expected` is a tuple in `assertTensorhenTuplesAlmostEqual`, check if the `actual` has the same size before iterating the elements.

It should also check if `actual` is a tuple, but the existing code has heavily abused this func to compare `tensor` vs `tuple`, `list` vs `tuple`. Have to leave the fix to the future.

Reviewed By: skim9

Differential Revision: D46307992

fbshipit-source-id: f9a5a027f1e188396274b17890078ad54e6bbb8f