replace np.float with float [2]
Summary:
`numpy.float` has been deprecated and removed in numpy-1.20.0. It was an alias to the builtin `float` and should be safe to replace with `float`.

The whole change was mechanical with the following one-liner:
```
fbgr -sl '\bnp\.float\b' | xargs perl -pi -e 's,\bnp\.float\b,float,g'
```

Differential Revision: D46609274

fbshipit-source-id: 69397882b8a3a3097840bf6a9e341730479e3833