Remove the selected dim in _select_targets
Summary:
Remove the dummy dims of 1 of the output of `_select_targets` when the target is selected with `gather` , so it acts the same as when `target` is `int` or scalar.

Also add comments to explain `_select_targets`

(the function is way too confusing. We should seriously rethink what to support or suggest users to no longer use `target`)

Reviewed By: vivekmig

Differential Revision: D50619740

fbshipit-source-id: b1325adcfb6f95b521c6eb889a18f1e72e05b9b5