Correct default `method` argument for captum.attr.visualization.visualize_t… (#1152)
Summary:
The default argument for `method` in `captum.attr.visualization.visualize_timeseries_attr` is currently `"individual_channels"`, which is not a valid option, resulting in an exception if used. This PR changes the default method to `"overlay_individual"`, which is what the docs indicate the default should be.

Pull Request resolved: https://github.com/pytorch/captum/pull/1152

Reviewed By: aobo-y

Differential Revision: D47197267

Pulled By: vivekmig

fbshipit-source-id: 4d87f792b742fafbb9c30e84247c830e93df1187