mps fix (#1227)
Summary:
By adding that single conversion from float64 to float32 the integrated gradients is fully compatible with MPS backend, which it wasn't previously. The change can be accepted as valid for any backend since float64 is a precision almost nobody using nowadays in deep learning. It seems that the default for torch.tensor is not updated with current trends.

Pull Request resolved: https://github.com/pytorch/captum/pull/1227

Reviewed By: cyrjano

Differential Revision: D54047349

Pulled By: vivekmig

fbshipit-source-id: 1ffda83f065a3f14fa3c5b0229fe0feb5035cc99