Split the get_random_model_and_data() method [1/n] (#1252)
Summary:
Pull Request resolved: https://github.com/pytorch/captum/pull/1252

As titled. The get_random_model_and_data() method is used to construct testing data for influence and it is reported as too complex by flake8 (https://www.flake8rules.com/rules/C901.html). This series of diff will split the method and abstract the common parts.

This diff isolate the model part for different gpu usage settings. It also eliminate the mix usage of bool and str.

Reviewed By: yucu

Differential Revision: D55165054

fbshipit-source-id: ee58b34c8f4ef7a3a93478a0665ad0f30070a832