Fix the tests with exceptions assertions having multiple top-level statements (#1264)
Summary:

As titled. Add error message checks as well for test comprehensity and code readability.

B908: Contexts with exceptions assertions like with self.assertRaises or with pytest.raises should not have multiple top-level statements. Each statement should be in its own context. That way, the test ensures that the exception is raised only in the exact statement where you expect it.

Differential Revision: D55344319